Java Programming
Create a superclass MathException and two subclasses OverFlowException and
UnderFlowException from it. Write code segment that instantiate an integer
variable and initialize it with some value. If the value is greater that 100, an
OverFlowException is thrown, otherwise if the value is less that 0 an
UnderFlowException is thrown. Now, write three catch blocks for
MathException, OverFlowException and for UnderFlowException. Write
appropriate message in the catch blocks and observe the result. In the first catch block re-throw the exception caught. In the other two catch
blocks write appropriate statements to handle it.